# packages/sdk/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "c-plane API",
    "description": "Control plane API for managing projects, environments, containers, and databases",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/organization/{organization_id}/projects/{project_id}/registry/repositories/{repository_id}/tags": {
      "get": {
        "tags": [
          "registry"
        ],
        "operationId": "list_tags",
        "parameters": [
          {
            "name": "organization_id",
            "in": "path",
            "description": "Organization ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "project_id",
            "in": "path",
            "description": "Project ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "repository_id",
            "in": "path",
            "description": "Registry repository ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Repository tag names",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepositoryTagsResponse"
                }
              }
            }
          },
          "403": {
            "description": "Organization access required"
          },
          "404": {
            "description": "Project or registry repository not found"
          },
          "503": {
            "description": "Container registry is unavailable"
          }
        },
        "security": [
          {
            "apiKey": []
          }
        ],
        "x-cplane-required-scope": "registry:read"
      }
    },
    "/api/organization/{organization_id}/projects/{project_id}/registry/repositories/{repository_id}/tags/{tag}": {
      "delete": {
        "tags": [
          "registry"
        ],
        "operationId": "delete_tag",
        "parameters": [
          {
            "name": "organization_id",
            "in": "path",
            "description": "Organization ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "project_id",
            "in": "path",
            "description": "Project ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "repository_id",
            "in": "path",
            "description": "Registry repository ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "tag",
            "in": "path",
            "description": "Tag name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Tag deleted"
          },
          "403": {
            "description": "Organization access required"
          },
          "404": {
            "description": "Project, registry repository, or tag not found"
          },
          "503": {
            "description": "Container registry is unavailable"
          }
        },
        "security": [
          {
            "apiKey": []
          }
        ],
        "x-cplane-required-scope": "registry:delete"
      }
    },
...
    "/api/organization/{organization_id}/registry": {
      "get": {
        "tags": [
          "registry"
        ],
        "operationId": "get_registry",
        "parameters": [
          {
            "name": "organization_id",
            "in": "path",
            "description": "Organization ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Managed Registry configuration",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManagedRegistryResponse"
                }
              }
            }
          },
          "404": {
            "description": "Managed Registry is not activated"
          }
        },
        "security": [
          {
            "apiKey": []
          }
        ],
        "x-cplane-required-scope": "registry:read"
      },
      "put": {
        "tags": [
          "registry"
        ],
        "operationId": "activate_registry",
        "parameters": [
          {
            "name": "organization_id",
            "in": "path",
            "description": "Organization ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActivateManagedRegistryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Managed Registry already active",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManagedRegistryResponse"
                }
              }
            }
          },
          "201": {
            "description": "Managed Registry activated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManagedRegistryResponse"
                }
              }
            }
          },
          "404": {
            "description": "Organization or active region not found"
          },
          "409": {
            "description": "Active region has no S3 provider"
          }
        },
        "security": [
          {
            "apiKey": []
          }
        ],
        "x-cplane-required-scope": "registry:create"
      }
    },
...
    "/internal/organizations/{organization_id}/registry": {
      "get": {
        "tags": [
          "internal"
        ],
        "operationId": "resolve_registry",
        "parameters": [
          {
            "name": "organization_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResolvedManagedRegistry"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "serviceToken": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ActivateManagedRegistryRequest": {
        "type": "object",
        "required": [
          "region_id"
        ],
        "properties": {
          "region_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "details": {},
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "ManagedRegistryResponse": {
        "type": "object",
        "required": [
          "organization_id",
          "region_id",
          "status",
          "storage_revision",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "string"
          },
          "organization_id": {
            "type": "string",
            "format": "uuid"
          },
          "region_id": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string"
          },
          "storage_revision": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "RepositoryTagsResponse": {
        "type": "object",
        "required": [
          "tags"
        ],
        "properties": {
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "ResolvedManagedRegistry": {
        "type": "object",
        "required": [
          "organization_id",
          "organization_slug",
          "storage_revision",
          "status",
          "access_key_id",
          "bucket_name",
          "storage_endpoint_url"
        ],
        "properties": {
          "access_key_id": {
            "type": "string"
          },
          "bucket_name": {
            "type": "string"
          },
          "organization_id": {
            "type": "string",
            "format": "uuid"
          },
          "organization_slug": {
            "type": "string"
          },
          "repository_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "repository_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "string"
          },
          "storage_endpoint_url": {
            "type": "string"
          },
          "storage_revision": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
    },
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API key with the operation's required scope."
      },
      "registryBasic": {
        "type": "http",
        "scheme": "basic"
      },
      "serviceToken": {
        "type": "apiKey",
        "in": "header",
        "name": "x-cplane-token",
        "description": "Internal service token."
      }
    }
  },
...
}